+++ pyplotformat/plot/format.py
@@ -148,9 +148,16 @@ class Format():
 
         # Set tick formatting
         # =========================================================================================
-        ticks_loc = self.axes.get_xticks().tolist()
+        if kwargs["x_tick_loc"] is None:
+            ticks_loc = self.axes.get_xticks().tolist()
+        else:
+            ticks_loc = kwargs["x_tick_loc"]
         self.axes.xaxis.set_major_locator(mticker.FixedLocator(ticks_loc))
-        ticks_loc = self.axes.get_yticks().tolist()
+
+        if kwargs["y_tick_loc"] is None:
+            ticks_loc = self.axes.get_yticks().tolist()
+        else:
+            ticks_loc = kwargs["y_tcik_loc"]
         self.axes.yaxis.set_major_locator(mticker.FixedLocator(ticks_loc))
 
         self.axes.set_xticklabels(self.axes.get_xticks(), **self.tickfont)