+++ pyplotformat/io/__init__.py
@@ -4,3 +4,4 @@ This subpackage provides utility functions for the pyPlotFormat package.
 
 from .save import save_figure, load_figure
 from .write import write_pdf
+from .inkscaper import inkscape